# src/NPaperless.Services/wwwroot/openapi-original.json
{
  "openapi" : "3.0.1",
  "info" : {
    "title" : "Paperless Rest Server",
    "version" : "1.0"
  },
  "servers" : [ {
    "url" : "/"
  } ],
  "paths" : {
...
    "/api/documents/" : {
      "get" : {
        "operationId" : "GetDocuments",
        "parameters" : [ {
          "explode" : true,
          "in" : "query",
          "name" : "Page",
          "required" : false,
          "schema" : {
            "format" : "int32",
            "type" : "integer"
          },
          "style" : "form"
        }, {
          "explode" : true,
          "in" : "query",
          "name" : "page_size",
          "required" : false,
          "schema" : {
            "format" : "int32",
            "type" : "integer"
          },
          "style" : "form"
        }, {
          "explode" : true,
          "in" : "query",
          "name" : "query",
          "required" : false,
          "schema" : {
            "type" : "string"
          },
          "style" : "form"
        }, {
          "explode" : true,
          "in" : "query",
          "name" : "ordering",
          "required" : false,
          "schema" : {
            "type" : "string"
          },
          "style" : "form"
        }, {
          "explode" : true,
          "in" : "query",
          "name" : "tags__id__all",
          "required" : false,
          "schema" : {
            "items" : {
              "format" : "int32",
              "type" : "integer"
            },
            "type" : "array"
          },
          "style" : "form"
        }, {
          "explode" : true,
          "in" : "query",
          "name" : "document_type__id",
          "required" : false,
          "schema" : {
            "format" : "int32",
            "type" : "integer"
          },
          "style" : "form"
        }, {
          "explode" : true,
          "in" : "query",
          "name" : "storage_path__id__in",
          "required" : false,
          "schema" : {
            "format" : "int32",
            "type" : "integer"
          },
          "style" : "form"
        }, {
          "explode" : true,
          "in" : "query",
          "name" : "correspondent__id",
          "required" : false,
          "schema" : {
            "format" : "int32",
            "type" : "integer"
          },
          "style" : "form"
        }, {
          "explode" : true,
          "in" : "query",
          "name" : "truncate_content",
          "required" : false,
          "schema" : {
            "type" : "boolean"
          },
          "style" : "form"
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "examples" : {
                  "default" : {
                    "value" : "{\"count\":1,\"next\":null,\"previous\":null,\"all\":[1],\"results\":[{\"id\":1,\"correspondent\":1,\"document_type\":1,\"storage_path\":1,\"title\":\"test\",\"content\":\"test\",\"tags\":[1],\"created\":\"2023-09-08T00:00:00Z\",\"created_date\":\"2023-09-08\",\"modified\":\"2023-09-08T12:15:24.652098Z\",\"added\":\"2023-09-08T12:05:28.338799Z\",\"archive_serial_number\":1,\"original_file_name\":\"test.pdf\",\"archived_file_name\":\"2023-09-08 edtest test.pdf\",\"owner\":3,\"user_can_change\":true,\"notes\":[{\"id\":1,\"note\":\"pdf in testdir\",\"created\":\"2023-09-08T12:11:05.848131Z\",\"document\":1,\"user\":3},{\"id\":2,\"note\":\"pdf 2 in test\",\"created\":\"2023-09-08T12:15:12.328015Z\",\"document\":1,\"user\":3}]}]}"
                  }
                },
                "schema" : {
                  "$ref" : "#/components/schemas/GetDocuments_200_response"
                }
              }
            },
            "description" : "Success"
          }
        },
        "tags" : [ "Documents" ]
      }
    },
...
    "/api/documents/post_document/" : {
      "post" : {
        "operationId" : "UploadDocument",
        "requestBody" : {
          "content" : {
            "multipart/form-data" : {
              "encoding" : {
                "title" : {
                  "style" : "form"
                },
                "created" : {
                  "style" : "form"
                },
                "document_type" : {
                  "style" : "form"
                },
                "tags" : {
                  "style" : "form"
                },
                "correspondent" : {
                  "style" : "form"
                },
                "document" : {
                  "style" : "form"
                }
              },
              "schema" : {
                "$ref" : "#/components/schemas/UploadDocument_request"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "Success"
          }
        },
        "tags" : [ "Documents" ]
      }
    },
    "/api/documents/bulk_edit/" : {
      "post" : {
        "operationId" : "BulkEdit",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "examples" : {
                "default" : {
                  "value" : "{\"documents\":[2],\"method\":\"redo_ocr\",\"parameters\":{}}"
                }
              },
              "schema" : {
                "$ref" : "#/components/schemas/BulkEdit_request"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "Success"
          }
        },
        "tags" : [ "Documents" ]
      }
    },
    "/api/documents/selection_data/" : {
      "post" : {
        "operationId" : "SelectionData",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "examples" : {
                "default" : {
                  "value" : "{\"documents\":[]}"
                }
              },
              "schema" : {
                "$ref" : "#/components/schemas/SelectionData_request"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "examples" : {
                  "default" : {
                    "value" : "{\"selected_correspondents\":[{\"id\":1,\"document_count\":0}],\"selected_tags\":[{\"id\":1,\"document_count\":0}],\"selected_document_types\":[{\"id\":1,\"document_count\":0}],\"selected_storage_paths\":[{\"id\":1,\"document_count\":0}]}"
                  }
                },
                "schema" : {
                  "$ref" : "#/components/schemas/SelectionData_200_response"
                }
              }
            },
            "description" : "Success"
          }
        },
        "tags" : [ "Documents" ]
      }
    },
    "/api/documents/{id}/suggestions/" : {
      "get" : {
        "operationId" : "GetDocumentSuggestions",
        "parameters" : [ {
          "explode" : false,
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "int32",
            "type" : "integer"
          },
          "style" : "simple"
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "examples" : {
                  "default" : {
                    "value" : "{\"correspondents\":[],\"tags\":[],\"document_types\":[],\"storage_paths\":[],\"dates\":[]}"
                  }
                },
                "schema" : {
                  "$ref" : "#/components/schemas/GetDocumentSuggestions_200_response"
                }
              }
            },
            "description" : "Success"
          }
        },
        "tags" : [ "Documents" ]
      }
    },
    "/api/documents/{id}/thumb/" : {
      "get" : {
        "operationId" : "GetDocumentThumb",
        "parameters" : [ {
          "explode" : false,
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "int32",
            "type" : "integer"
          },
          "style" : "simple"
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/pdf" : {
                "schema" : {
                  "format" : "binary",
                  "type" : "string"
                }
              }
            },
            "description" : "Success"
          }
        },
        "tags" : [ "Documents" ]
      }
    },
    "/api/documents/{id}/download/" : {
      "get" : {
        "operationId" : "DownloadDocument",
        "parameters" : [ {
          "explode" : false,
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "int32",
            "type" : "integer"
          },
          "style" : "simple"
        }, {
          "explode" : true,
          "in" : "query",
          "name" : "original",
          "required" : false,
          "schema" : {
            "type" : "boolean"
          },
          "style" : "form"
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/pdf" : {
                "schema" : {
                  "format" : "binary",
                  "type" : "string"
                }
              }
            },
            "description" : "Success"
          }
        },
        "tags" : [ "Documents" ]
      }
    },
    "/api/documents/{id}/preview/" : {
      "get" : {
        "operationId" : "GetDocumentPreview",
        "parameters" : [ {
          "explode" : false,
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "int32",
            "type" : "integer"
          },
          "style" : "simple"
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/pdf" : {
                "schema" : {
                  "format" : "binary",
                  "type" : "string"
                }
              }
            },
            "description" : "Success"
          }
        },
        "tags" : [ "Documents" ]
      }
    },
    "/api/documents/{id}/metadata/" : {
      "get" : {
        "operationId" : "GetDocumentMetadata",
        "parameters" : [ {
          "explode" : false,
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "int32",
            "type" : "integer"
          },
          "style" : "simple"
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "examples" : {
                  "default" : {
                    "value" : "{\"original_checksum\":\"39b80972dc7f1e975d9d2e8b92ce4e8b\",\"original_size\":11868,\"original_mime_type\":\"application/pdf\",\"media_filename\":\"testdirp.pdf\",\"has_archive_version\":true,\"original_metadata\":[],\"archive_checksum\":\"3597d60204a4b200b97e728622f704ab\",\"archive_media_filename\":\"testdirp.pdf\",\"original_filename\":\"test.pdf\",\"lang\":\"et\",\"archive_size\":15561,\"archive_metadata\":[{\"namespace\":\"http://ns.adobe.com/pdf/1.3/\",\"prefix\":\"pdf\",\"key\":\"Producer\",\"value\":\"pikepdf 7.2.0\"},{\"namespace\":\"http://ns.adobe.com/xap/1.0/\",\"prefix\":\"xmp\",\"key\":\"ModifyDate\",\"value\":\"2023-09-08T12:04:59+00:00\"},{\"namespace\":\"http://ns.adobe.com/xap/1.0/\",\"prefix\":\"xmp\",\"key\":\"CreateDate\",\"value\":\"2023-09-08T11:46:26+00:00\"},{\"namespace\":\"http://ns.adobe.com/xap/1.0/\",\"prefix\":\"xmp\",\"key\":\"CreatorTool\",\"value\":\"ocrmypdf 14.3.0 / Tesseract OCR-PDF 5.3.0\"},{\"namespace\":\"http://ns.adobe.com/xap/1.0/mm/\",\"prefix\":\"xmpMM\",\"key\":\"DocumentID\",\"value\":\"uuid:7af613b2-865c-11f9-0000-925a0e2aa608\"},{\"namespace\":\"http://purl.org/dc/elements/1.1/\",\"prefix\":\"dc\",\"key\":\"format\",\"value\":\"application/pdf\"},{\"namespace\":\"http://purl.org/dc/elements/1.1/\",\"prefix\":\"dc\",\"key\":\"title\",\"value\":\"Ohne Titel\"},{\"namespace\":\"http://www.aiim.org/pdfa/ns/id/\",\"prefix\":\"pdfaid\",\"key\":\"part\",\"value\":\"2\"},{\"namespace\":\"http://www.aiim.org/pdfa/ns/id/\",\"prefix\":\"pdfaid\",\"key\":\"conformance\",\"value\":\"B\"},{\"namespace\":\"http://purl.org/dc/elements/1.1/\",\"prefix\":\"dc\",\"key\":\"creator\",\"value\":\"None\"},{\"namespace\":\"http://ns.adobe.com/xap/1.0/\",\"prefix\":\"xmp\",\"key\":\"MetadataDate\",\"value\":\"2023-09-08T12:04:59.067651+00:00\"}]}"
                  }
                },
                "schema" : {
                  "$ref" : "#/components/schemas/GetDocumentMetadata_200_response"
                }
              }
            },
            "description" : "Success"
          }
        },
        "tags" : [ "Documents" ]
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "GetDocuments_200_response" : {
        "example" : {
          "next" : 6,
          "all" : [ 5, 5 ],
          "previous" : 1,
          "count" : 0,
          "results" : [ {
            "owner" : 4,
            "user_can_change" : true,
            "archive_serial_number" : 2,
            "notes" : [ {
              "note" : "note",
              "created" : "created",
              "document" : 1,
              "id" : 7,
              "user" : 1
            }, {
              "note" : "note",
              "created" : "created",
              "document" : 1,
              "id" : 7,
              "user" : 1
            } ],
            "added" : "added",
            "created" : "created",
            "title" : "title",
            "content" : "content",
            "tags" : [ 3, 3 ],
            "storage_path" : 9,
            "archived_file_name" : "archived_file_name",
            "modified" : "modified",
            "correspondent" : 2,
            "original_file_name" : "original_file_name",
            "id" : 5,
            "created_date" : "created_date",
            "document_type" : 7
          }, {
            "owner" : 4,
            "user_can_change" : true,
            "archive_serial_number" : 2,
            "notes" : [ {
              "note" : "note",
              "created" : "created",
              "document" : 1,
              "id" : 7,
              "user" : 1
            }, {
              "note" : "note",
              "created" : "created",
              "document" : 1,
              "id" : 7,
              "user" : 1
            } ],
            "added" : "added",
            "created" : "created",
            "title" : "title",
            "content" : "content",
            "tags" : [ 3, 3 ],
            "storage_path" : 9,
            "archived_file_name" : "archived_file_name",
            "modified" : "modified",
            "correspondent" : 2,
            "original_file_name" : "original_file_name",
            "id" : 5,
            "created_date" : "created_date",
            "document_type" : 7
          } ]
        },
        "properties" : {
          "count" : {
            "type" : "integer"
          },
          "next" : {
            "type" : "integer"
          },
          "previous" : {
            "type" : "integer"
          },
          "all" : {
            "items" : {
              "type" : "integer"
            },
            "type" : "array"
          },
          "results" : {
            "items" : {
              "$ref" : "#/components/schemas/GetDocuments_200_response_results_inner"
            },
            "type" : "array"
          }
        },
        "required" : [ "all", "count", "next", "previous", "results" ],
        "type" : "object"
      },
...
      "UploadDocument_request" : {
        "properties" : {
          "title" : {
            "type" : "string"
          },
          "created" : {
            "format" : "date-time",
            "type" : "string"
          },
          "document_type" : {
            "format" : "int32",
            "type" : "integer"
          },
          "tags" : {
            "items" : {
              "format" : "int32",
              "type" : "integer"
            },
            "type" : "array"
          },
          "correspondent" : {
            "format" : "int32",
            "type" : "integer"
          },
          "document" : {
            "items" : {
              "format" : "binary",
              "type" : "string"
            },
            "type" : "array"
          }
        },
        "type" : "object"
      },
      "BulkEdit_request" : {
        "properties" : {
          "documents" : {
            "items" : {
              "type" : "integer"
            },
            "type" : "array"
          },
          "method" : {
            "type" : "string"
          },
          "parameters" : {
            "type" : "object"
          }
        },
        "required" : [ "documents", "method", "parameters" ],
        "type" : "object"
      },
...
    }
  }
}